Fix issue #7 in truncate EDF annotations
This fixes issue #7 by loading time__resolution into the truncation routine and using it to calculate total time of raw signal to a value truncated to a whole number of epochs, and using that to calculate needed truncation of the EDF annotations.
Tested working with 10s and 5s time_resolution on all data in Bear_selection.